read/write global vars in zmemory instead of map
There are opcodes that can read and write any address in dynamic
memory, so they can read/write global vars without specifically asking
to. This means my way of storing them in a map for ease of access could
lead to weird behavior in any game that decided to do that.